+++ src/main/java/com/thoughtworks/codepairing/model/Product.java
@@ -1,17 +1,19 @@
 package com.thoughtworks.codepairing.model;
 
+import java.math.BigDecimal;
+
 public class Product {
-    private final double price;
+    private final BigDecimal price;
     private final String productCode;
     private final String name;
 
-    public Product(double price, String productCode, String name) {
+    public Product(BigDecimal price, String productCode, String name) {
         this.price = price;
         this.productCode = productCode;
         this.name = name;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
